mainから具体的な処理を省くリファクタリング
