+++ typeScriptScreepsProject/src/main.ts
@@ -1,8 +1,7 @@
 import { roleBuilder } from './modules/role/role.builder';
-import { managementCreepCount, minimumHarvesterCount } from './managementCreep';
+import { managementCreepCount } from './managementCreep';
 import { spawnUpgrader } from './modules/spawn/spawn.upgrader';
 import { logger } from './modules/logger';
-import { findTarget } from './modules/findTarget';
 import { spawnRepaierer } from './modules/spawn/spawn.repaierer';
 import { deleteDeadCreepMemory } from './modules/deleteDeadCreepMemory';
 import { roleHarvester } from './modules/role/role.harvester';
@@ -14,6 +13,8 @@ import { findCreepsByRole } from './modules/find/findCreepsByRole';
 import { spawnTransporter } from './modules/spawn/spawn.transporter';
 import { roleTransporter } from './modules/role/role.transporter';
 import { spawnerStatus } from './modules/spawnerStatus';
+import { roleHealer } from './modules/role/roleHealer';
+import { roleAttacker } from './modules/role/roleAttacker';
 
 export interface CreepMemory {
   [key: string]: any;
@@ -80,12 +81,7 @@ module.exports.loop = function () {
     const creep = Game.creeps[name];
     switch ((creep.memory as CreepMemory).role) {
       case 'harvester': {
-        const harvestingTargets = findTarget.filling(creep);
-        if (harvestingTargets.length === 0) {
-          roleUpgrader.run(creep);
-        } else {
-          roleHarvester.run(creep);
-        }
+        roleHarvester.run(creep);
         break;
       }
 
@@ -100,59 +96,22 @@ module.exports.loop = function () {
       }
 
       case 'builder': {
-        const buildingTargets = creep.room.find(FIND_CONSTRUCTION_SITES);
-        const repaierTargets = creep.room.find(FIND_STRUCTURES, {
-          filter: (object) => object.hits < object.hitsMax,
-        });
-        if (harvesters.length < minimumHarvesterCount) {
-          roleHarvester.run(creep);
-        } else if (buildingTargets.length > 0) {
-          roleBuilder.run(creep);
-        } else if (repaierTargets.length > 0) {
-          roleRepaierer.run(creep);
-        } else {
-          roleUpgrader.run(creep);
-        }
+        roleBuilder.run(creep);
         break;
       }
 
       case 'repaierer': {
-        const repaierTargets = creep.room.find(FIND_STRUCTURES, {
-          filter: (object) => object.hits < object.hitsMax,
-        });
-        if (harvesters.length < minimumHarvesterCount) {
-          roleHarvester.run(creep);
-        } else if (repaierTargets.length > 0) {
-          roleRepaierer.run(creep);
-        } else {
-          roleUpgrader.run(creep);
-        }
+        roleRepaierer.run(creep);
         break;
       }
 
       case 'attacker': {
-        const targets = creep.room.find(FIND_HOSTILE_CREEPS);
-        if (targets.length) {
-          if (creep.attack(targets[0]) == ERR_NOT_IN_RANGE) {
-            creep.moveTo(targets[0], {
-              visualizePathStyle: { stroke: '#ff0000' },
-            });
-          }
-        }
+        roleAttacker(creep);
         break;
       }
 
       case 'healer': {
-        // TODO: SourceのIDを指定して、同時採掘可能な分のCreep数だけ割り当てる
-        // TODO: Mapから、Sourceを取得し、同時採掘可能なCreep数を割り出す
-        const targets = creep.room.find(FIND_HOSTILE_CREEPS);
-        if (targets.length) {
-          if (creep.attack(targets[0]) == ERR_NOT_IN_RANGE) {
-            creep.moveTo(targets[0], {
-              visualizePathStyle: { stroke: '#ff0000' },
-            });
-          }
-        }
+        roleHealer(creep);
         break;
       }
     }
