+++ typeScriptScreepsProject/src/modules/role/role.harvester.ts
@@ -4,15 +4,27 @@ import { memoryManager } from '../memoryManager';
 import { CreepMemory } from '../../main';
 import { isCreepStoreEmpty, isCreepStoreFull } from '../check/check.store';
 import { actionMove } from '../action/action.move';
+import { roleUpgrader } from './role.upgrader';
 
 const isHarvesting = (creep: Creep) => {
   return (creep.memory as CreepMemory).refueling;
+};
+
+const settingRole = (creep: Creep) => {
+  const harvestingTargets = findTarget.filling(creep);
+  if (harvestingTargets.length === 0) {
+    roleUpgrader.run(creep);
+  } else {
+    return;
+  }
 };
 
 // TODO: harvesterがharvestする場所を固定にする
 // TODO: harvestTargetIdをMemoryに保存する処理をaction.harvest.tsに移動する
 export const roleHarvester = {
   run: function (creep: Creep) {
+    settingRole(creep);
+
     (creep.memory as CreepMemory).roleAs = 'harvester';
 
     // TODO: Creepが作りたての状態を考慮できていないため、roleのみが設定された状態のCreepの扱いを決める
